refactor(temporal): separate worker deployment schemas
Why: the rollout module exceeded the repository's 500-line lint limit after the rollout safety proofs landed.\n\nWhat: move deployment and version parsing schemas into a focused shared module without changing behavior.\n\nVerification: bun --no-install --bun vitest --config ../../vitest.config.ts run src/lib/worker-deployment-rollout.test.ts; bunx turbo run typecheck lint --filter=@shepherdjerred/temporal
